Move function from utils to venv-pip and refactor tests
