+++ tests/v2/06_tasks/test_tasks_v2_utils.py
@@ -1,8 +1,6 @@
 import pytest
 
-from fractal_server.tasks.v2.utils import _init_venv_v2
 from fractal_server.tasks.v2.utils import get_python_interpreter_v2
-from tests.execute_command import execute_command
 
 
 async def test_get_python_interpreter_v2(
@@ -41,34 +39,3 @@ async def test_get_python_interpreter_v2(
     assert python39 == MOCK_PYTHON_3_9
     python312 = get_python_interpreter_v2("3.12")
     assert python312 == MOCK_PYTHON_3_12
-
-
-@pytest.mark.parametrize("python_version", [None, "3.10"])
-async def test_init_venv(tmp_path, python_version):
-    """
-    GIVEN a path and a python version
-    WHEN _init_venv_v2() is called
-    THEN a python venv is initialised at path
-    """
-    venv_path = tmp_path / "fractal_test"
-    venv_path.mkdir(exist_ok=True, parents=True)
-    logger_name = "fractal"
-
-    try:
-        python_bin = await _init_venv_v2(
-            path=venv_path,
-            logger_name=logger_name,
-            python_version=python_version,
-        )
-    except ValueError as e:
-        pytest.xfail(reason=str(e))
-
-    assert venv_path.exists()
-    assert (venv_path / "venv").exists()
-    assert (venv_path / "venv/bin/python").exists()
-    assert (venv_path / "venv/bin/pip").exists()
-    assert python_bin.exists()
-    assert python_bin == venv_path / "venv/bin/python"
-    if python_version:
-        version = await execute_command(f"{python_bin} --version")
-        assert python_version in version
